Time: 0 ms (100%), Space: 28.2 MB (8.33%) - LeetHub

+++ 3402-minimum-operations-to-make-columns-strictly-increasing/3402-minimum-operations-to-make-columns-strictly-increasing.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int minimumOperations(vector<vector<int>>& grid) {
-        long ans = 0;
-        for(long c = 0; c < grid[0].size(); ++c){
+        long long int ans = 0;
+        for(long long int c = 0; c < grid[0].size(); ++c){
             for(int r = 1; r < grid.size(); ++r){
                 if(grid[r][c] > grid[r-1][c]) continue;
                 ans += grid[r-1][c] + 1 - grid[r][c];
